Creating website elements
This is the very first commit for a project, for which I have been
working on for nearly a month. Currently some of elements for the
Landing Page are initially designed, keeping it simple at first.
There are just few more to add; those are: FAQ, Offer showcase, Footer
and Team Showcase (currently working on it).

This project is currently not supposed to be a full website project, but
rather a landing page design mixed with fancy scroll-based animation and
exceptional design patterns. Maybe in the future more Tab views will be
added, however this is not the main point of the project, at least for
now.