Complete initial creation of the components
Another step has been made to the build of the project. All the
components shown on the landing Page has just been prepared. That still
does not mean they are fully prepared though. Some of them needs some
tweaks and changes, while others are completely fine-tuned to the
style of the website. The very next update will focus exactly on
remaking the components which are required to be in this way.