Complete Reviews Section
The Reviews section is finally completed ! This took a lot of time
actually, mostly because there was a bug in the TeamShowcase section,
which started to cause some major issues with GSAP animations, which
were not always lined up with their intended start and end positions.
Since it's repaired, the Reviews part works just fine now.