Complete Collab section
Finally, the new section to the website is done! This was the toughest
part to remake so far. It's because the old section "Company in numbers"
got completely replaced with the "Collab" part - built almost
entirely from scratch. Also this very section introduces a brand new
scroll mechanic into the website, giving fresh, exeptional UX.