Finish the Newsletter section
The very last section has finally been completed! At this point, the
crucial design parts are done, the only left thing in this matter is to
fix some minor flaws about them, which should not take much time. Apart,
the next step is to fill some website sections with a content (that is
mostly a dummy text right now) - this means both text and images.