Completed  ContactShowcase component
The very last component has just been added to the project! All the
animations for most of the page seems to be working and integrating
well with each other (at least for desktops). The only significant
corrections are need for the very first sections - Intro and
OfferShowcase. Those two are fairly behind in terms of design and
overall integrity with other parts of the site. So that is goning to be
the next goal for now.