Partially update OfferShowcase component
OfferShowcase simply needed some visual update, and there it is! The
most of the job here is done - with amazing neon-like visual for each
offer, some styling corrections and more. Last part to focus on inside
this component is the "See full offer" mini-section and the end
transition - and with those things to be done, the final improvements
are closer and closer to complete.