Complete designing the page!
After countless amounts of work, the page is now fully designed! All
the parts are in place, each designed and (almost) prepared to be
deployed further to production.

Now the current goal is going to be slightly different - website needs
some actual content, not just a placeholder text. And that will be the
main focus for now. Also, some components will receive some slight
tweaks along the way!