Add content to the website
The Skylite website is now fueled with a content that has an actual
meaning and that is a cohesive part of the project. Took longer than
expected, but the end effect look good! Getting really close to work
on the very last feature - mobile queries as a cherry on top for this
fun project.