Another corrections & quality updates
This is the very last design part across the whole development stage
of this project. The website looks so much smmoother now, with all the
animations and transitions in place, every single detail works as
intended. This almost finishes the project - the remain part is to
prepare the project for different viewport sizes.