Complete portrait mobiles design
The project is now fully implemented for mobile devices! It has been a
quite tough step to go through, but at the end of the day turned out
to be successful. The mobile layout still looks very closely to the
original design, because there were just a plenty elements on the whole
page that were essential to remove completely. That being said, most
page segment preserved all of the content as for huger screens, with
just some important alignment changes added.

Nevertheless, tablets need some tweaks. Maybe not that huge as for
mobile devices, but still some essential layout-wise upgrades should be
prepared.