Complete landscape tablets layout
Landscape tablets are finished. The approach here was a bit different
than as for mobile devices. It is because instead of adding another
media query, some actual fixes were added for the base styling, so that
landscape tablet screens would also fall for this style path. Also,
slightly changed media query max-width / max-height value to a slightly
lower ones.