Complete media queries for tablests
Tablet devices are pretty much designed to the most of what can be done
without using component-specific media queries directly for tablets.
The only media query that is used for tablets specifically, is a global
one that manages the proper font size across the webpage. Other than
that, tablet devices gotten some limits in terms of max sizing of
some text elements, so that they will not stretch long across one line,
but rather be divided into more separate lines.