Bugfix solves for mobile devices & more
Solved a very annoying issue on mobile devices with address bar - when
triggering browser's native address bar to show or hide, a weird
random auto scroll behavior was happening. It is no longer a problem
now!

Slighlty changed offer items background colors for portrait devices.
The cool gradient move effect, while worked great for landscape devices,
was not any good for portrait ones. Insted of the animation, a stable,
solid background is applied, and the whole section became much more
visible.

Last but not least - got rid of one of the diagonal lines background
layer, which simply looked bad on mobile devices. This got replaced for
the good-old, dotted background layer, which was still in use for
quite some time.